requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search #tool that helps in web search
fastapi[all]
uvicorn
groq
openai